rework aflfast
